requirements.txt
opencv-python
numpy
scikit-learn
Pillow
colorama
pywin32